requirements.txt
# requirements.txt
pandas==2.2.2
pyodbc==5.1.0
python-dotenv==1.0.1
pyadomd==0.1.1
pythonnet==3.0.3
SQLAlchemy==2.0.29
numpy==1.26.4
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0
